fix darwin build by moving code
Fixes issue #135
